--- .pre-commit-config.yaml
@@ -1,5 +1,0 @@
-repos:
--   repo: https://github.com/ambv/black
-    rev: stable
-    hooks:
-    - id: black
