+++ graphene_boilerplate/gql/item/__init__.py
@@ -1,5 +1,5 @@
 import graphene
-from graphene_boilerplate.gql.item.fields import ItemSchema
+from graphene_boilerplate.gql.item.schema import ItemSchema
 from graphene_boilerplate.gql.item.mutations import CreateItemMutation
 from graphene_boilerplate.gql.item.queries import resolve_item
 
@@ -9,7 +9,7 @@ class ItemQuery(graphene.ObjectType):
         type=ItemSchema,
         id_=graphene.Int(required=True),
         resolver=resolve_item,
-        description="Get item by id",
+        description="Get item by id_",
     )
