+++ graphene_boilerplate/gql/item/schema/__init__.py
@@ -1,12 +1,12 @@
 import graphene
 from graphene import relay
 from graphene_sqlalchemy import SQLAlchemyObjectType
-from graphene_boilerplate import Item
-from graphene_boilerplate.user_schema import UserSchema
+from graphene_boilerplate.model import Item
+from graphene_boilerplate.gql.user.schema import UserSchema
 
 
 class ItemSchema(SQLAlchemyObjectType):
-    id_ = graphene.Int(description="id of item")
+    id_ = graphene.Int(description="id_ of item")
     key = graphene.String(description="item key")
     value = graphene.JSONString(description="dictionary on the item object")
     user = graphene.Field(UserSchema, description="user who owns the item")
