+++ graphene_boilerplate/gql/user/__init__.py
@@ -1,6 +1,6 @@
 import graphene
 from graphene_boilerplate.gql.user.queries import resolve_user
-from graphene_boilerplate.user_schema import UserSchema
+from graphene_boilerplate.gql.user.schema import UserSchema
 
 
 class UserQuery(graphene.ObjectType):
